Properly handle export_file targets in haskell rules
Summary:
## Problem
The `LoadCache.hs` module on the `//haxl/core/tests:haxl-core-unit-lib` rule depends on a text file, `LoadCache.txt`. The text file is "exposed" via an `expose_file` rule that works in buck1, but not in buck2.

## Why?
FWIU, the `LoadCache.hs` haskell module depends on `LoadCache.txt` **as a source file**. The `export_file` rule will make the file available as a resource, i.e. in the `buck-out` directory, which wasn't enough in this case.

## Proposed solution
Create a new provider for `export_file` which lets dependant rules know when a dependency is one that type and handle the exported file accordingly.

In the case of `LoadCache.hs`, we'll add it as a hidden dependency to the compilation step.

Reviewed By: pepeiborra

Differential Revision: D46274259

fbshipit-source-id: 94a9d7a36adf3c4e72c71841283ea792e5fc43cf