Extract get_packages_info to use in haskell_ghci
Summary:
## This stack
See summary of D46476480.

## RFC
I'm exposing the option to specify the version in `-expose-package` because that's done in the sigma_ghc_flags script (D46480501).
Would it be OK to just always specify the version?
I didn't do that right away because it would modify the behaviour of `_compile` in `haskell.bzl`, which, IIUC, is used in all the other rules...

## This diff
Extracts out the function that gets the exposed packages, package dbs and all the haskell transitive dependencies, so I can use it in the haskell_ghci rule in D46480501 and D46486127.

Reviewed By: simonmar

Differential Revision: D46275685

fbshipit-source-id: 7fb55b090573a30f583669f067686273ca19ebe0