Rename argsfile_args to args
Summary:
`argsfile_args` (with shell quoting) was named as such to coexist with `args` (which did not contain shell quoting).

With D43707948, `args` is no longer all args, but the `args_without_file_prefix_args` (renamed in the previous diff).

As such, rename this field, as it is a better name.

Reviewed By: milend

Differential Revision: D46762974

fbshipit-source-id: c1d7aeba58e3c236e39cb11a1bc349e2eafe6855