Return a CompileArgsfiles containing relative and absolute argsfiles for Swift
Summary:
In order to cleanly merge argsfiles in the Cxx layer, we need to provide relative and absolute argsfiles for use.

At various points in this stack, we added support for `CompileArgsfiles` and integrated it in. We will replace `CxxAdditionalArgsfileParams` with `CompileArgsfiles` and in order to do so, let's first set them in Swift.

Note: The `swiftc` compiler in Xcode correctly handles relative paths for indexing, so we don't need to generate absolute versions. We just have to pass a `CompileArgsfiles.absolute` for API clarity later.

Reviewed By: milend

Differential Revision: D46743448

fbshipit-source-id: 637b230b3a1ad14b736e0f3e915f4890097bc85c