Move method_id definitions to a hrl file
Summary:
# Context
In this stack we want to make cth_tpx handle correctly the case where it runs together with user hooks; in particular we want to detect failures in any hook callback. This will require introducing a couple more modules split the work currently done in cth_tpx.

# Problem
The type `method_id()` and the macros with the possible values is used everywhere, and once we introduce other modules, they will need to be aware of it.

# This diff
We move the type and constants to a header that can later be shared among them. In doing this, we notice that the type `method_id()` was wrongly defined (as in, it doesn't match the actual code), which was quite confusing. We fix this here as well

Reviewed By: TheGeorge

Differential Revision: D46690284

fbshipit-source-id: 755abaa56c4d2c83dab357783d4508e629555e86