buck2: add python version check in rustc_action.py
Summary:
Small quality of life improvement.

This script uses `asyncio.run` which was added in Python 3.7.

My machine was misconfigured and was using Python 3.6 by default, have a nicer error message for this case.

This script is also used in OSS, should be an improvement for external users as well.

Reviewed By: zertosh

Differential Revision: D47100792

fbshipit-source-id: 99d576bbc8fbf555ee244fa3731328f711ed7179