Opt-out prebuilt C/C++ libs from dist LTO processing
Summary:
In practice, prebuilt C/C++ libraries always contain native code and so
don't need to go through the same dist LTO processing that built-from-source
objects do.

This diff adds a new field to `ArchiveLinkable` to signal to the dist LTO
machinery that an archive doesn't contain bitcode, and and can avoid
the LTO processing overhead.

This also unblocks D46799896, which migrates `prebuilt_cxx_library_group`
to use `Linkable`s, and causes an issue w/ use-cases of this rule using
`--start-group`/`--end-group`, which doesn't mix with dist LTO's use of
`--start-lib`/`--end-lib` (which `ld.lld` doesn't allow to be nested).

Reviewed By: luciang

Differential Revision: D47336408

fbshipit-source-id: 35aa5d557cd5c2f2770894bb9f36736c16e98fe9