Export split debug handling (take 2)
Summary:
When using split-debug modes and LTO on macos or linux, the link step will emit
debug info objects that we need to track via external build info.

Export this directory in the `LinkedObject` struct, so that we can access it from
the caller.

Second attempt fixes a breakage w/ fbobjc builds.

Reviewed By: athmasagar

Differential Revision: D47353926

fbshipit-source-id: 1fb72332a47ec2e263a0f6dd9440f2701ecbd481