Fix handling of erlang:exit/1,2
Summary:
# Problem 1
Let's say we had a test like this:
```
test_foo(_Config) ->
    Parent = self(),
    spawn(fun() -> erlang:exit(Parent, "die die die" end),
    ok.
```

Running this from the shell would show an error saying that there was a formatting error:

```
✗ blah_SUITE - test_foo
unable to format failure reason, please report.
{error,{test_isany,{sentinel_crash,"die die die"}}}
```

# Problem 2
Worse, if we had something like:
```
test_bar(_Config) ->
    ct:fail("fail, baby, fail").
```

we'd get a bad crash:

```
✗ blah_SUITE - test_bar
failed with error:
** exception error: no function clause matching ct_daemon_printer:'-chop_stack/1-fun-0-'(102) (buck2/prelude/erlang/common_test/test_exec/src/ct_daemon_printer.erl, line 84)
     in function  lists:takewhile_1/2 (lists.erl, line 1499)
     in call from ct_daemon_printer:print_error/4 (buck2/prelude/erlang/common_test/test_exec/src/ct_daemon_printer.erl, line 61)
     in call from test:'-do_plain_test_run/1-fun-2-'/3 (buck2/prelude/erlang/common_test/test_cli_lib/src/test.erl, line 360)
     in call from maps:fold_1/3 (maps.erl, line 411)
     in call from test:do_plain_test_run/1 (buck2/prelude/erlang/common_test/test_cli_lib/src/test.erl, line 358)
```

Under the hood, `ct:fail/1` just calls `erlang:exit/1`.

# This diff
The source for Problem 2 was in `ct_daemon_hooks:format_ct_error/3`, where we were assuming that an `exit` would have no stacktrace, but this is incorrect. The case where we don't have a stacktrace is when the process is killed by `erlang:exit/2`, but in that case, the execution never makes it to `ct_daemon_hooks:format_ct_error/3`.

Because `ct:fail("foo")` is essentially `exit({test_case_failure, "foo"})`, what would happen is that later on `"foo"` would get confused with a stack trace (since `is_list("foo")` passes)

For Problem 1, we add a special case in the formatter

Reviewed By: TheGeorge

Differential Revision: D47402434

fbshipit-source-id: 1073a90155746cdcea19da12a1cc494fe2538bc4