Unify LTO+split-debug output dir handling across linux and darwin
Summary:
Currently, for linux, we rely on users setting the special linker flags to specify a
split-debug output to use at link time (so that LTO can output debug info there),
and use a pretty gross hack to inject the output object we generate (see next diff).

This diff adds these linker flags into the rules, and unifies this with how darwin
does it.

Reviewed By: shayne-fletcher

Differential Revision: D47474599

fbshipit-source-id: dd8aa35ce6c0a52687e56de7ae2e9f133fb25cf1