Support anonymous C++ links for link groups
Summary:
This adds an anonymous target layer which serializes/deserializes
the `cxx_link` API across an `ctx.actions.anon_target()` to allow
(mostly) seamless anonymous linking in existing prelude code.

The feature is exposed via a new `anonymous` parameter to `cxx_link`
and the C++ link groups code is updated to use it.

I went back and forth on the best way to implement this and finally
settled on this approach -- largely maintaining the `cxx_link` API
and doing the heavy lifting in trampoline code to do serialization
and deserialization w/ minor refactoring to facilitate in the stack --
to make it trivial to deploy in the prelude.

That said, I do think we probably need to do some more cleanup/
refactoring of the `cxx_link` function to make it easier to re-use
in situations like this (e.g. separating out the logic that assembles
the `LinkedObject` result from the logic creates the actions to
generate the raw results).

#forceTDHashing

Reviewed By: christycylee

Differential Revision: D46691318

fbshipit-source-id: c3114ef816de6b86f9949548ca37c9241edbc74f