Use correct category for stripping actions
Summary:
**Problem**

We use `strip_shared_lib` as a category for the strip action for both shared libraries and binaries. This can be misleading when debugging and inspecting logs.

**Solution**

If available, use the category suffix provided by the caller. For binaries, the category prefix is `executable`, so we get the correct categorisation.

Reviewed By: IanChilds, krallin

Differential Revision: D47546238

fbshipit-source-id: d1b22220ea859583d0dec54249b10a6467dbf7e3