Implement swift_toolchain support_cxx_interoperability_mode
Summary:
Implementation for Buck2 of the same functionality implemented in Buck1 in
D47634402.

With Xcode 15 betas we have seen a new Swift compiler flag
--cxx-compatibility-mode instead of the legacy
--enable-experimental-cxx-interop (which is not longer supported).

Luckily Buck2 does not use the system Xcode so everything is driven by xbat
specs that were modified in D47578295 (and D47578296) to provide the needed
information.

Reviewed By: rmaz

Differential Revision: D47643194

fbshipit-source-id: c95692caa59dbf33f7a7729283d651bd2c08541d