Enable static linking libpython.a
Summary:
IG wants to link against `libpython.a` statically, but because references to
`libpython.so` is pervasive in the codebase, even after adding `libpython.so`
will eventually be linked in. Even though the static linker will resolve symbols
to the static version, the binary still end up with a `DT_NEEDED` tag to
third-party `libpython.so` instead of platform `libpython.so`, this creates
a load order conflict at runtime. This diff adds the buck option
`-c fbcode.link_static_libpython3.10` to drop `libpython.so` from the link graph
and add `libpython.a` to the link graph.

Reviewed By: itamaro

Differential Revision: D47306409

fbshipit-source-id: 2998e4a7d066dbf6f2ef216905f3ab416ca2c0ea