buck2: http_archive don't *expect* that the daemon digest config allows SHA1
Summary:
Like it says in the title. This is a fix for a Github issue that's been open
for a little while where OSS users get forced into those actions running local
only because they don't provide a SHA1 (and if they did they might get the
wrong behavior here).

Note: I made this prefer_local because that a) makes more sense and b)
we have a test that sets the config to blake3-only and also passes `--remote-only`,
predictably that blows up if we use `local_only = True`.

Fixes: https://github.com/facebook/buck2/pull/303

Reviewed By: themarwhal

Differential Revision: D47665044

fbshipit-source-id: 762b621f80d621ea9f264426e5a645cdae454f77